+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 -   Added autocomplete support. Try `symbol-bootstrap autocomplete` and follow the instructions (Thanks @44uk).
 -   Renamed `supernode` keywords for `rewardProgram` for clarification.  Supernode is a type of Reward Program.
 -   Renamed command from `enrolSupernode` for `enrolRewardProgram`. 
+-   Added preset configurable `connectionPoolSize` to the Rest Gateway configuration.
 
 ## [0.4.2] - Feb-2-2021